Models Refactor
