+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/CodeActions/Models/RazorCodeActionResolutionParams.cs
@@ -2,13 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using MediatR;
-using Newtonsoft.Json.Linq;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer
+namespace Microsoft.AspNetCore.Razor.LanguageServer.CodeActions.Models
 {
     internal class RazorCodeActionResolutionParams : IRequest<RazorCodeActionResolutionResponse>
     {
         public string Action { get; set; }
-        public JObject Data { get; set; }
+        public object Data { get; set; }
     }
 }
